tests with Estevao and Ulysses

+++ control.py
@@ -8,8 +8,8 @@ Created on Sat May 23 17:55:13 2015
 def control(error):
     signal = 0
     try:    
-        p = 1/float(1000)
-        i = 1/float(10000)
+        p = 1/float(100)
+        i = 1/float(100000)
         
         # If there is a change of signal, reset
 #        if ((error[-2] >= 0) and (error[-1] < 0)) or ((error[-2] < 0) and (error[-1] >= 0)):
@@ -26,6 +26,8 @@ def control(error):
         elif signal < 0:
             signal = 0
             error[-1] = 0
+
+        # return 1
             
     except ValueError:
         return "Control error"
